# optimize memplanner