@nimlgen 2024-08-07
Idea is not bad, will reopen when it'll be ready to merge